pyproject.toml
[tool.poetry]
name = "töröö"
version = "0.1.0"
description = ""
authors = ["Julian Leucker <leuckerj@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.1.1"
numpy = "^1.22.4"
noise = "^1.2.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
icecream = "^2.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
